python/requirements.txt
requests==2.27.1
mysqlclient==1.4.6
python-dotenv==0.18.0